Add files via upload
Abstractive summarizer for IBM. The encoder-decoder model uses mostly Amazon text/summary data to train on but has a couple hundred text/summary samples from the CFPB dataset mixed in (at x20 weight, to compensate for their lack of volume). The model is regularized with 0.25% dropout (regular and recurrent). The output summarizations can be viewed against the ten test samples at the  bottom of the notebook.